Refactored getCourseGroupRepos to getCourseGroup and more
Getting various types of repositories is done a bit differently
in several places; this commit is a first step to reorganize these
methods so that it becomes more consistent and easier to reason
about.
